Refactor to more manageable class sizes
