+++ src/lib/db/feature-strategy-store.ts
@@ -301,7 +301,7 @@ class FeatureStrategiesStore {
     async getFeatures(
         featureQuery?: IFeatureToggleQuery,
         archived: boolean = false,
-        includeStrategyId: boolean = true,
+        isAdmin: boolean = true,
     ): Promise<IFeatureToggleClient[]> {
         const environments = [':global:'];
         if (featureQuery?.environment) {
@@ -369,9 +369,7 @@ class FeatureStrategiesStore {
                 feature.strategies = [];
             }
             if (r.strategy_name) {
-                feature.strategies.push(
-                    this.getAdminStrategy(r, includeStrategyId),
-                );
+                feature.strategies.push(this.getAdminStrategy(r, isAdmin));
             }
             if (feature.enabled === undefined) {
                 feature.enabled = r.enabled;
@@ -384,65 +382,15 @@ class FeatureStrategiesStore {
             feature.stale = r.stale;
             feature.type = r.type;
             feature.variants = r.variants;
-            feature.lastSeenAt = r.last_seen_at;
             feature.project = r.project;
-            feature.createdAt = r.created_at;
+            if (isAdmin) {
+                feature.lastSeenAt = r.last_seen_at;
+                feature.createdAt = r.created_at;
+            }
             acc[r.name] = feature;
             return acc;
         }, {});
         return Object.values(featureToggles);
-    }
-
-    async getProjectOverview(
-        projectId: string,
-        archived: boolean = false,
-    ): Promise<IFeatureOverview[]> {
-        const rows = await this.db('features')
-            .where({ project: projectId, archived })
-            .select(
-                'features.name as feature_name',
-                'features.type as type',
-                'features.created_at as created_at',
-                'features.last_seen_at as last_seen_at',
-                'features.stale as stale',
-                'feature_environments.enabled as enabled',
-                'feature_environments.environment as environment',
-                'environments.display_name as display_name',
-            )
-            .fullOuterJoin(
-                'feature_environments',
-                'feature_environments.feature_name',
-                'features.name',
-            )
-            .fullOuterJoin(
-                'environments',
-                'feature_environments.environment',
-                'environments.name',
-            );
-        if (rows.length > 0) {
-            const overview = rows.reduce((acc, r) => {
-                if (acc[r.feature_name] !== undefined) {
-                    acc[r.feature_name].environments.push(
-                        this.getEnvironment(r),
-                    );
-                } else {
-                    acc[r.feature_name] = {
-                        type: r.type,
-                        name: r.feature_name,
-                        createdAt: r.created_at,
-                        lastSeenAt: r.last_seen_at,
-                        stale: r.stale,
-                        environments: [this.getEnvironment(r)],
-                    };
-                }
-                return acc;
-            }, {});
-            return Object.values(overview).map((o: IFeatureOverview) => ({
-                ...o,
-                environments: o.environments.filter(f => f.name),
-            }));
-        }
-        throw new NotFoundError(`Could not find project with id ${projectId}`);
     }
 
     async getStrategyById(id: string): Promise<IFeatureStrategy> {
@@ -465,8 +413,6 @@ class FeatureStrategiesStore {
             .onConflict(['environment', 'feature_name'])
             .merge('enabled');
     }
-
-    async;
 
     async enableEnvironmentForFeature(
         feature_name: string,
@@ -513,33 +459,6 @@ class FeatureStrategiesStore {
             .update(update)
             .returning('*');
         return mapRow(row[0]);
-    }
-
-    async getMembers(projectId: string): Promise<number> {
-        const rolesFromProject = this.db('role_permission')
-            .select('role_id')
-            .distinct()
-            .where({ project: projectId });
-
-        const numbers = await this.db('role_user')
-            .distinct()
-            .count('user_id as members')
-            .whereIn('role_id', rolesFromProject)
-            .first();
-        const { members } = numbers;
-        if (typeof members === 'string') {
-            return parseInt(members, 10);
-        }
-        return members;
-    }
-
-    // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-    private getEnvironment(r: any): IEnvironmentOverview {
-        return {
-            name: r.environment,
-            displayName: r.display_name,
-            enabled: r.enabled,
-        };
     }
 
     private getAdminStrategy(
