+++ src/lib/types/services.ts
@@ -21,6 +21,7 @@ import UserFeedbackService from '../services/user-feedback-service';
 import FeatureToggleServiceV2 from '../services/feature-toggle-service-v2';
 import EnvironmentService from '../services/environment-service';
 import FeatureTagService from '../services/feature-tag-service';
+import ProjectHealthService from '../services/project-health-service';
 
 export interface IUnleashServices {
     accessService: AccessService;
@@ -46,4 +47,5 @@ export interface IUnleashServices {
     userService: UserService;
     versionService: VersionService;
     featureTagService: FeatureTagService;
+    projectHealthService: ProjectHealthService;
 }
