+++ src/test/e2e/services/environment-service.test.ts
@@ -71,9 +71,7 @@ test('Can connect environment to project', async () => {
         variants: [],
     });
     await service.connectProjectToEnvironment('test-connection', 'default');
-    const overview = await stores.featureStrategiesStore.getProjectOverview(
-        'default',
-    );
+    const overview = await stores.projectStore.getProjectOverview('default');
     overview.forEach(f => {
         expect(f.environments).toEqual([
             {
@@ -92,9 +90,7 @@ test('Can remove environment from project', async () => {
     });
     await service.removeEnvironmentFromProject('test-connection', 'default');
     await service.connectProjectToEnvironment('removal-test', 'default');
-    let overview = await stores.featureStrategiesStore.getProjectOverview(
-        'default',
-    );
+    let overview = await stores.projectStore.getProjectOverview('default');
     expect(overview.length).toBeGreaterThan(0);
     overview.forEach(f => {
         expect(f.environments).toEqual([
@@ -106,9 +102,7 @@ test('Can remove environment from project', async () => {
         ]);
     });
     await service.removeEnvironmentFromProject('removal-test', 'default');
-    overview = await stores.featureStrategiesStore.getProjectOverview(
-        'default',
-    );
+    overview = await stores.projectStore.getProjectOverview('default');
     expect(overview.length).toBeGreaterThan(0);
     overview.forEach(o => {
         expect(o.environments).toEqual([]);
